Fix clippy --all-targets warnings and check all targets in CI
Fixes existing clippy errors and warnings that only appeared with
--all-targets, including PI-approximation literals in tests, an absurd
extreme i64::MAX comparison, len without is_empty, field_reassign, and
io_other_error. Also updates the justfile lint/format recipes to pass
--all-targets so test, bench, and example clippy issues are caught in
CI going forward.

+++ compiler/benchmarks/benches/st_benchmark.rs
@@ -131,7 +131,7 @@ fn bench_branching(c: &mut Criterion) {
             BenchmarkId::from_parameter(branches),
             &container,
             |bufs| {
-                bufs.vars[0] = Slot::from_i32((branches - 1) as i32);
+                bufs.vars[0] = Slot::from_i32(branches - 1);
             }
         );
     }
